+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Better handling of lists.
 * Fix the "blahtex" math engine on 1.8.7.
 * "script" and "style" tags now have their generated CDATA tags escaped so the scripts/styles actually work. #120
+* CDATA escaping is only added to "script" and "style" tags if it is necessary, rather than all the time.
 * "section" tags are now equivalent to "div" tags - they won't get wrapped in a "p" tag and they can have markdown processing turned on with a "markdown" attribute. #117
 * In some situations where Maruku might detect invalid Markdown and omit it from the output, it will now simply pass through the input as if it were regular text. Discovered in #124.
 * HTML like <script> tags will no longer cause errors if they don't have a newline separating them from other text. #123